Move all pytorch imports to estimate.py
